Improve file structure inside the frontend. Add frontend/backend directory. Initialize spring application. Refactor RegisterScreen.tsx, LoginScreen.tsx. Separate their style in a different .ts file. Add file structure to the backend. Add entity - UserAcccount.java. Add enum - UserType.java
