+++ backend/src/main/java/com/cafe/backend/BackendApplication.java
@@ -1,0 +1,13 @@
+package com.cafe.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackendApplication.class, args);
+	}
+
+}
