+++ backend/src/main/java/com/cafe/backend/UserType.java
@@ -1,0 +1,8 @@
+package com.cafe.backend;
+
+public enum UserType {
+    ADMIN,
+    CUSTOMER,
+    CLIENT,
+    ANONYMOUS
+}
