+++ frontend/app/screens/Register/Register.style.ts
@@ -1,0 +1,21 @@
+import {StyleSheet} from "react-native";
+
+export default StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: '#EDC4B3',
+    },
+    title: {
+        fontSize: 24,
+        fontWeight: 'bold',
+        marginBottom: 20,
+        color: '#774936',
+    },
+    button: {
+        marginTop: 10,
+        width: '90%',
+        backgroundColor: '#774936',
+    },
+});
