+++ frontend/app/screens/Register/RegisterScreen.tsx
@@ -3,7 +3,8 @@ import { View, StyleSheet, Text } from 'react-native';
 import { Button } from 'react-native-paper';
 import { useNavigation } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../navigation/navigation';
+import { RootStackParamList } from '../../navigation/navigation';
+import styles from "./Register.style";
 
 const RegisterScreen = () => {
   const navigation = useNavigation<StackNavigationProp<RootStackParamList>>();
@@ -22,25 +23,5 @@ const RegisterScreen = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#EDC4B3',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    color: '#774936',
-  },
-  button: {
-    marginTop: 10,
-    width: '90%',
-    backgroundColor: '#774936',
-  },
-});
 
 export default RegisterScreen;
